% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{hroc}
\alias{hroc}
\title{Hierarchical Reciever Operating Characteristic}
\usage{
hroc(dat, labelID, methods = c("single", "complete", "average"),
  clustObj = NULL, ...)
}
\arguments{
\item{dat}{An NxP data matrix where the P columns are going to be clustered.}

\item{labelID}{A list of length P giving a vector of integer or text
labels for each column that was clustered.  Vectors of labels are
accepted to allow observations to belong to multiple categories.  In this
case, pairs of observations will be considered linked if any of the
labels in the vector match any of the labels corresponding to the other
observation.  NA's are allowed which results in the ROC plot being made
based on the clustering of the full data but only using the non-NA labels
in computing sensitivity and specificity.}

\item{methods}{A vector of strings that determine what sort of linkage
function should be used in the hierarchical clustering.  Parameters
can be passed to the agnes function through "..." but this is separate
because it allows users to put multiple HROC curves on the same plot.
By default single, complete and average linkage functions will be
plotted together.}
}
\description{
Function for creating HROC plots.  These plots show the reciever operating
characteristic for a given set of labels (possibly incomplete) and a type
of hierarchical clustering algorithm.  The distance at which any two
points are merged along a dengrogram is used as a predictor of linkage
while true pair-wise linkage is based on the labels.
Hierarchical clustering is done via the \code{\link{agnes}} function from
the \code{\link{cluster}} package.  All of the agnes function parameters
can be passed through the hroc function. In addition to creating a plot,
this function returns a table with all of the pairwise combinations,
along with their respective linkages and distances.
}
