% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexFunctions.R
\name{pairwise}
\alias{pairwise}
\title{Creating Pairwise indices}
\usage{
pairwise(trueLabels, clustLabels)
}
\arguments{
\item{trueLabels}{Either a character or numeric vector denoting the known
or hypothesized labels for each observation.}

\item{clustLabels}{Either a character or numeric vector denoting the
clustering assignments for each observation.}
}
\description{
Function for creating indices based on the 2x2 table of pairwise
comparisons. This includes sensitivity, specificity, positive predictive
value, negative predictive value and a kappa statistic. The function
returns a list object containing all of these pariwise indices.
}
