% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexFunctions.R
\name{triplets}
\alias{triplets}
\title{Creating indices based on triplets of observations}
\usage{
triplets(trueLabels, clustLabels)
}
\arguments{
\item{trueLabels}{Either a character or numeric vector denoting the known
or hypothesized labels for each observation.}

\item{clustLabels}{Either a character or numeric vector denoting the
clustering assignments for each observation.}
}
\description{
Function for creating indices based on the 5x5 table of triplet
comparisons. This includes kappa computed on the 5x5 table, along with
five other indices representing submatrix associations.
}
